Filter runaway cron jobs
A very rare bug where a cron job that doesn't delete itself causes the
same Tweet to be sent repeatedly. If that Tweet has media attached, it
can end up being sent and accepted by Twitter dozens of times because
the URL for media uploads might vary. Currently in testing.